% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineModels.R
\name{combineModels}
\alias{combineModels}
\title{Combine outputs}
\usage{
combineModels(...)
}
\arguments{
\item{...}{One or more output objects, to be combined to list of models.}
}
\description{
This function takes model objects (class \code{outputs}) of JJM and generate an object 
with combined models.
}
\examples{
\dontrun{
mod1 <- runJJM(modelName = "mod2.1")
mod2 <- runJJM(modelName = "mod2.2")
mod3 <- runJJM(modelName = "mod2.3")

mod_123 = combineModels(mod1, mod2, mod3)
}
}
