% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{compareModels}
\alias{compareModels}
\title{Compare combined JJM outputs}
\usage{
compareModels(mods)
}
\description{
This function takes a vector of model names, reads in the JJM runs, and combines them.
Basically a wrapper function for \code{combineModels}.
Assumes model runs are in the same folder.
}
\examples{
\dontrun{

mod_123 = compareModels(c("h1_0.00", "h1_0.01", "h1_0.02")
}
}
