% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjmr-main.R
\name{diagnostics}
\alias{diagnostics}
\title{Generate Assessment plots from single model}
\usage{
diagnostics(object, ...)
}
\arguments{
\item{object}{Object ob class outputs.}

\item{...}{Extra arguments}
}
\description{
Function to generate plots from results of readJJM function
}
\examples{
model = readJJM(modelName = "mod2.4")
diagnostics(object = model)
}
