% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_bmsy.R
\name{fixed_bmsy}
\alias{fixed_bmsy}
\title{Calculate or input a fixed Bmsy value for the jjm model
Updates the msy_mt table in the jjm output with new B/Bmsy ratios.}
\usage{
fixed_bmsy(mod, refpt = NULL)
}
\arguments{
\item{mod}{jjm object that is a list of lists}

\item{refpt}{A number to input as Bmsy. If not filled, calculated as the average of the Bmsy estimated for the last ten years (as determined in SCW14 benchmark 2022)}
}
\value{
A model
}
\description{
Calculate or input a fixed Bmsy value for the jjm model
Updates the msy_mt table in the jjm output with new B/Bmsy ratios.
}
\examples{
# fixed_bmsy(mod_h1_1.00, refpt=5500) # To input a fixed Bmsy
# fixed_bmsy(mod_h1_1.00) # To calculate the Bmsy
}
