% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_fits.R
\name{get_age_fits}
\alias{get_age_fits}
\title{Get fits to indices}
\usage{
get_age_fits(models)
}
\arguments{
\item{models}{an object of class jjm.output}
}
\value{
a tidy dataframe of age fits
}
\description{
Get fits to indices
}
