% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catchabilities.R
\name{get_catchabilities}
\alias{get_catchabilities}
\title{Get estimated catchability coefficients}
\usage{
get_catchabilities(models)
}
\arguments{
\item{models}{an object of class jjm.output}
}
\value{
a data frame of estimated catchabilities
}
\description{
Get estimated catchability coefficients
}
