% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mortality.R
\name{get_fishing_mortality}
\alias{get_fishing_mortality}
\title{Get fishing mortality at age}
\usage{
get_fishing_mortality(models)
}
\arguments{
\item{models}{an object of class jjm.output}
}
\value{
a tidy dataframe of fishing mortality at age
}
\description{
Get fishing mortality at age
}
