% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_fits.R
\name{get_index_fits}
\alias{get_index_fits}
\title{Get fits to indices}
\usage{
get_index_fits(models)
}
\arguments{
\item{models}{an object of class jjm.output}
}
\value{
a tidy dataframe of index fits
}
\description{
Get fits to indices
}
