% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_len_fits.R
\name{get_len_fits}
\alias{get_len_fits}
\title{Get fits to length compositions}
\usage{
get_len_fits(models)
}
\arguments{
\item{models}{an object of class jjm.output}
}
\value{
a tidy dataframe of length composition fits
}
\description{
Get fits to length compositions
}
