% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msy_mt.R
\name{get_msy_mt}
\alias{get_msy_mt}
\title{Get and tidy msy_my table}
\usage{
get_msy_mt(models)
}
\arguments{
\item{models}{series of class jjm.output}
}
\value{
a tidy msy_mt
}
\description{
Get and tidy msy_my table
}
\examples{
\dontrun{

mod0.00 <- readJJM("h2_0.00", path = "config", input = "input")
get_msy_mt(mod0.00)
}

}
