% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recruits.R
\name{get_recruits}
\alias{get_recruits}
\title{Get estimated recruits}
\usage{
get_recruits(models)
}
\arguments{
\item{models}{an object of class jjm.output}
}
\value{
a tidy dataframe of recruits
}
\description{
Get estimated recruits
}
