% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selectivities.R
\name{get_selectivities}
\alias{get_selectivities}
\title{Get and tidy selectivity-at-age ogives over time by model and fleet}
\usage{
get_selectivities(models)
}
\arguments{
\item{models}{an object of class jjm.output}
}
\value{
a tidy data frame of selectivity estimates
}
\description{
Get and tidy selectivity-at-age ogives over time by model and fleet
}
\examples{
\dontrun{
h1.mod <- jjmR::readJJM("h2_0.02", path = "config", input = "input")
selectivities <- get_selectivities(h1.mod)
}

}
