% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_totals.R
\name{get_totals}
\alias{get_totals}
\title{Get total metrics (biomass, spawning biomass, and recruitment)}
\usage{
get_totals(models)
}
\arguments{
\item{models}{an object of class jjm.output}
}
\value{
a dataframe of total values
}
\description{
Get total metrics (biomass, spawning biomass, and recruitment)
}
