% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{geth}
\alias{geth}
\title{Add hypothesis number to a model name}
\usage{
geth(mod, h = hyp)
}
\arguments{
\item{mod}{A character string of a model name.}

\item{h}{A character string containing the hypothesis to use.}
}
\value{
A character string containing the hypothesis name and the model name.
}
\description{
Add hypothesis number to a model name
}
\examples{
geth("1.00", "h1")
}
