% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{getter}
\alias{getter}
\title{Get elements of a list by string matching on name}
\usage{
getter(x, pattern = "^sel_", things = NA)
}
\arguments{
\item{x}{the object}

\item{pattern}{the string pattern (regex supported) to search for}
}
\value{
an object subset to matches in names with strings
}
\description{
Get elements of a list by string matching on name
}
