% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_selectivities.R
\name{plot_selectivities}
\alias{plot_selectivities}
\title{Plot selectivities by age, year, fleet, and model}
\usage{
plot_selectivities(
  sels,
  fleet = "fsh",
  alpha = 0.4,
  scale = 4,
  size = 0.5,
  years = "all"
)
}
\arguments{
\item{sels}{selectivity data frame generated by \code{get_selectivities}}

\item{fleet}{fleets to plot: "fsh" (fishery), "ind" (survey), or "all" (both)}
}
\value{
a ggplot2 plot object
}
\description{
Plot selectivities by age, year, fleet, and model
}
\examples{

\dontrun{

oldnewMods <- combineModels(mod0.00,mod_prev)
selectivities <- get_selectivities(oldnewMods)
plot_selectivities(selectivities)

}

}
