% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExFiles.R
\name{readExFiles}
\alias{readExFiles}
\title{Read external files}
\usage{
readExFiles(
  fileName,
  type,
  path = NULL,
  version = "2015MS",
  parameters = FALSE,
  parData,
  nameFishery,
  nameIndex,
  nAges,
  nStock = NULL
)
}
\arguments{
\item{fileName}{filename}

\item{type}{type}

\item{path}{path}

\item{version}{version of JJM, default to "2015MS" (2015 SC multi-stock).}

\item{parameters}{parameters}

\item{parData}{parData}

\item{nameFishery}{nameFishery}

\item{nameIndex}{nameIndex}

\item{nAges}{nAges}

\item{nStock}{nStock}
}
\description{
Read external files
}
