% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjmr-main.R
\name{readJJM}
\alias{readJJM}
\title{Read a model or list of models}
\usage{
readJJM(model, path = NULL, output = "results", input = NULL,
  version = "2015MS", ...)
}
\arguments{
\item{model}{String with the name of model that will be readed or run.}

\item{path}{Directory where the 'admb' folder is located.}

\item{...}{Extra arguments}

\item{modelName}{alias for \code{model} (to be deprecated).}
}
\description{
Function to read models and list if models and generate results
}
\examples{
readJJM(model = "mod2.4")
}

