% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report}
\alias{report}
\title{Create a report from JJM outputs}
\usage{
report(object, format, output, tangle = FALSE, tidy = TRUE, ...)
}
\arguments{
\item{object}{The object to create the report with, can be of classes 
'jjm.output' or 'jjm.diag' as created with \code{readJJM} or \code{diagnostics}.}

\item{format}{Format for the report: either "pdf", "html" or "word".}

\item{output}{Path to save the report, by default the working directory.}

\item{tangle}{Boolean, if TRUE the R script to create the report is produced.}

\item{tidy}{Boolean, if TRUE the intermediate files (Rmd, tex) are deleted.}

\item{...}{Extra arguments}
}
\description{
Function to create and save reports in PDF and MS Word formats.
}
\examples{
\dontrun{
report(mod0.0)
}
}
