% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{retro}
\alias{retro}
\title{Run a retrospective analysis diagnostic for a JJM model}
\usage{
retro(models, n = 5, output = "results", exec = NULL, iprint = 100,
  wait = TRUE, parallel = FALSE, temp = NULL, ...)
}
\arguments{
\item{models}{An object of class jjm.output}

\item{n}{Number of years to run a retrospective analysis.}

\item{output}{Path to save results.}

\item{exec}{Path to JJM executable file.}

\item{parallel}{Boolean flag to run models in parallel.}

\item{temp}{Folder to run retrospective analysis. If NULL, a temporal folder is used.}

\item{...}{Additional arguments passed to other functions.}
}
\description{
Run a retrospective analysis for a model
}
\examples{
readJJMConfig(mod1)
}

