% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjmr-main.R
\name{runit}
\alias{runit}
\title{Fit, run, read and plot a JJM model}
\usage{
runit(mod, est = FALSE, exec = NULL, path = "config", input = "input",
  output = "results", version = "2015MS", pdf = FALSE, portrait = TRUE)
}
\arguments{
\item{mod}{A character specifying the name of a model (by it's ctl filename).}

\item{est}{Boolean, should we run the parameter estimation for a model?}

\item{exec}{Path to the JJM executable file. By default, 'jjms' will be used.}

\item{path}{Directory where the configuration files will be written.}
}
\description{
Shortcut to fit, run, read and plot a JJM model
}
\examples{
writeJJM(mod1)
}
