% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jjm.R
\name{theme_jjm}
\alias{theme_jjm}
\title{ggplot2 theme for jjmR}
\usage{
theme_jjm(base_size = 14, ...)
}
\arguments{
\item{...}{}
}
\value{
a ggplot2 theme object
}
\description{
ggplot2 theme for jjmR
}
\examples{
 

ggplot(mtcars, aes(mpg)) + geom_histogram() + jjmr::theme_jjm()

}
