% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_jjm.R
\name{tidy_JJM}
\alias{tidy_JJM}
\title{Tidy results of JJM model}
\usage{
tidy_JJM(models)
}
\arguments{
\item{models}{an object of class jjm.output}
}
\value{
a list of tidy dataframes
}
\description{
Tidy results of JJM model
}
\examples{

\dontrun{
mod0.00 <- readJJM("h2_0.00", path = "config", input = "input")
tidy_jjm_results <- tidy_JJM(mod0.00)

}




}
