% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeJJM.R
\name{writeJJM}
\alias{writeJJM}
\title{Write dat and ctl files from a JJM model stored in R}
\usage{
writeJJM(object, path, ...)
}
\arguments{
\item{object}{An object of class jjm.config or jjm.output.}

\item{path}{Directory where the configuration files will be written.}

\item{...}{Additional arguments}
}
\description{
Function write to the disk dat and ctl files
}
\examples{
\dontrun{
writeJJM(mod1)
}
}
