% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjmr-main.R, R/jjmr-package.R
\docType{package}
\name{jjmR-package}
\alias{jjmR-package}
\alias{jjmR}
\title{Tools to process and get results from Joint Jack Mackerel (JJM) model outputs.}
\description{
Graphics and diagnostics tools for SPRFMO's Joint Jack Mackerel model.

Graphics and diagnostics tools for SPRFMO's Joint Jack Mackerel model.
}
\details{
\tabular{ll}{ Package: \tab jjmR\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2014-08-15\cr License: \tab TBD\cr }

\tabular{ll}{ Package: \tab jjmR\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2014-08-15\cr License: \tab TBD\cr }
}
\seealso{
Joint Jack Mackerel Repository <https://github.com/SPRFMO/jjm>

Joint Jack Mackerel Repository <https://github.com/SPRFMO/jjm>
}
\author{
Ricardo Oliveros-Ramos, Wencheng Lau-Medrano, Giancarlo Moron 
Josymar Torrejon and Niels Hintzen
}
\keyword{jjmr}
