% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJJM.R
\name{readJJM}
\alias{readJJM}
\title{Read a model or list of models}
\usage{
readJJM(
  model,
  path = NULL,
  output = "results",
  input = NULL,
  version = "2015MS",
  ...
)
}
\arguments{
\item{model}{String with the name of model that will be readed or run.}

\item{path}{Directory where the 'admb' folder is located.}

\item{output}{Path to the model outputs directory.}

\item{input}{Path to model inputs directory.}

\item{version}{version of JJM, default to "2015MS" (2015 SC multi-stock).}

\item{...}{Extra arguments}
}
\description{
Function to read models and list if models and generate results
}
\examples{
\dontrun{
readJJM(model = "mod2.4")
}
}
