% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJJMConfig.R
\name{readJJMConfig}
\alias{readJJMConfig}
\title{Read dat and ctl files from disk to create a jjm.config object.}
\usage{
readJJMConfig(model, path, input = NULL, ...)
}
\arguments{
\item{model}{Model object or outputs}

\item{path}{Path to the ctl file}

\item{input}{Path to the input files}

\item{...}{Additional arguments passed to other functions.}
}
\description{
Store in an R object (of class jjm.config) the dat and ctl files needed
to run a model.
}
\examples{
\dontrun{
readJJMConfig(mod1)
}
}
