% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{retro}
\alias{retro}
\title{Run a retrospective analysis diagnostic for a JJM model}
\usage{
retro(model, n = 5, output = "results", exec = NULL, parallel = FALSE,
  temp = NULL, wait = TRUE, iprint = 100, ...)
}
\arguments{
\item{model}{An object of class jjm.output}

\item{n}{Number of years to run a retrospective analysis.}

\item{output}{Path to save results.}

\item{exec}{Path to JJM executable file.}

\item{parallel}{Boolean flag to run models in parallel.}

\item{temp}{Folder to run retrospective analysis. If NULL, a temporal folder is used.}

\item{wait}{Boolen, passed to runJJM, should we wait for the parameter estimation?}

\item{iprint}{Command line argument passed to jjm.}

\item{...}{Additional arguments passed to other functions.}
}
\description{
Run a retrospective analysis for a model
}
\examples{
\dontrun{
readJJMConfig(mod1)
}
}
