% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runit.R
\name{runit}
\alias{runit}
\title{Fit, run, read and plot a JJM model}
\usage{
runit(
  mod,
  est = FALSE,
  exec = NULL,
  path = "config",
  input = "input",
  output = "results",
  version = "2015MS",
  pdf = FALSE,
  portrait = TRUE
)
}
\arguments{
\item{mod}{A character specifying the name of a model (by it's ctl filename).}

\item{est}{Boolean, should we run the parameter estimation for a model?}

\item{exec}{Path to the JJM executable file. By default, 'jjms' will be used.}

\item{path}{Directory where the configuration files will be written.}

\item{input}{Input}

\item{output}{Folder to save the outputs, 'arc' by default.}

\item{version}{version of JJM, default to "2015MS" (2015 SC multi-stock).}

\item{pdf}{Produce outputs in a pdf file?}

\item{portrait}{Orientation of the pdf output, default TRUE.}
}
\description{
Shortcut to fit, run, read and plot a JJM model
}
\examples{
\dontrun{
writeJJM(mod1)
}
}
